+++ Cultural Project/src/Objects/Entity.java
@@ -3,13 +3,13 @@ package Objects;
 import Graphics.Sprite;
 
 public class Entity {
-	protected int x, y;
+	protected float x, y;
 	protected char Chart;
 	protected int AnimN = 0;
 	protected Sprite sprite;
 	protected boolean IsMoving;
 
-	public Entity(int x, int y, char Chart) {
+	public Entity(float x, float y, char Chart) {
 		this.x = x;
 		this.y = y;
 		this.Chart = Chart;
